separate out checkbox component
